John Baskerville 8 and small code adjustments
Adding John Baskerville 8 which makes use of flex box and is a responsive. Small code adjustments have been made to files (tidying code to make it easier to read and some changes to make it more visually appealing.) CSS on has also been adjusted to make history of type responsive.